move context manager into a common place
