+++ src/plopp/graphics/static.py
@@ -2,7 +2,7 @@
 # Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
 
 from .fig import Figure
-from .io import fig_to_bytes
+from .utils import fig_to_bytes
 
 
 class StaticFig(Figure):
